runQueryWithRetry dedup
